COM-1201 add trello mixin
